Package removing was debugged.

git-svn-id: svn+ssh://svn.code.sf.net/p/dynamicdocs/code@285 05742d63-fc59-4ccd-b3f0-d7e2a2a8a88d

+++ ddocclient/kksfactory/KKSStuffFactory.cpp
@@ -798,7 +798,7 @@ void KKSStuffFactory :: delStuffElement (const QModelIndex& index, int entity, Q
             break;
         }
     }
-    int idStuffEnt = index.data (Qt::UserRole).toInt();
+    qint64 idStuffEnt = index.data (Qt::UserRole).toInt();
     int ier = oef->deleteOE (editor, io->id(), idStuffEnt, io->tableName(), stuffModel, index);
 /*    if (ier == OK_CODE && stuffModel)
     {
